+++ ARPickerInputViews/ARDatePickerInputView.m
@@ -26,6 +26,9 @@
 @implementation ARDatePickerInputView
 {
     UIToolbar *_toolbar;
+    UIBarButtonItem *_clearButton;
+    UIBarButtonItem *_flex;
+    UIBarButtonItem *_doneButton;
 }
 
 - (instancetype)initAsInputForTextField:(UITextField *)textField
@@ -46,10 +49,16 @@
         [toolbar sizeToFit];
         _toolbar = toolbar;
         
-        UIBarButtonItem *flex = [[UIBarButtonItem alloc] initWithBarButtonSystemItem:UIBarButtonSystemItemFlexibleSpace target:self action:nil];
-        UIBarButtonItem *rightButton = [[UIBarButtonItem alloc] initWithBarButtonSystemItem:UIBarButtonSystemItemDone target:self action:@selector(dismissDateOfBirthToolbar:)];
+        _clearButton = [[UIBarButtonItem alloc] initWithTitle:@"Clear" style:UIBarButtonItemStylePlain target:self action:@selector(clear:)];
+        _flex = [[UIBarButtonItem alloc] initWithBarButtonSystemItem:UIBarButtonSystemItemFlexibleSpace target:self action:nil];
+        _doneButton = [[UIBarButtonItem alloc] initWithBarButtonSystemItem:UIBarButtonSystemItemDone target:self action:@selector(dismissDateOfBirthToolbar:)];
         
-        toolbar.items = @[flex, rightButton];
+        if (_showClearButton) {
+            toolbar.items = @[_clearButton, _flex, _doneButton];
+        } else {
+            toolbar.items = @[_flex, _doneButton];
+        }
+        
         
         textField.inputView = self;
         textField.inputAccessoryView = toolbar;
@@ -107,6 +116,18 @@
     return _dateFormatter;
 }
 
+- (void)setShowClearButton:(BOOL)showClearButton
+{
+    if (_showClearButton != showClearButton) {
+        _showClearButton = showClearButton;
+        if (showClearButton) {
+            _toolbar.items = @[_clearButton, _flex, _doneButton];
+        } else {
+            _toolbar.items = @[_flex, _doneButton];
+        }
+    }
+}
+
 
 
 
@@ -145,6 +166,12 @@
 - (void)dismissDateOfBirthToolbar:(id)sender
 {
     [self.targetTextField resignFirstResponder];
+}
+
+- (void)clear:(id)sender
+{
+    self.targetTextField.text = nil;
+    [self dismissDateOfBirthToolbar:sender];
 }