+++ Demo/Demo/ViewController.h
@@ -23,7 +23,6 @@
 @property (nonatomic, weak, nullable) IBOutlet UITextField *listPickerTextField;
 
 
-
 @property (nonatomic, strong, nullable) ARDatePickerInputView *datePickerInputView;
 
 @property (nonatomic, strong, nullable) ARNumberPickerInputView *numberPickerInputView;