+++ CHANGES.md
@@ -1,3 +1,9 @@
+1.0.2 (13/07/2018)
+------------------
+* Added optional clear button to all the input views which will clear the target text fields.
+* Fixed crash when presenting an ARListPickerInputView with no content.
+
+
 1.0.1 (17/03/2017)
 ------------------
 * Setting the tint color now correctly tints the toolbar contents